Split GeBeOBot and context to different file
