+++ BotGeBeO.py
@@ -1,0 +1,32 @@
+from discord.ext import commands
+
+class GeBeOContext(commands.Context):
+    args_split = None
+    arg = None
+
+
+class BotGeBeO(commands.Bot):
+
+    config = {}
+
+    async def on_message(self, message):
+        ctx = await self.get_context(message, cls=GeBeOContext)
+        print(ctx.message.author.display_name + ": " + ctx.message.content)
+        ctx.args_split = ctx.message.content.split(' ')[1:]
+        ctx.arg = ' '.join(ctx.args_split)
+        await self.invoke(ctx)
+
+
+    async def on_command_error(self, ctx, error):
+        if isinstance(error, commands.errors.CommandInvokeError):
+            print(error.original)
+        elif isinstance(error, commands.errors.MissingPermissions):
+            await ctx.message.delete()
+            await ctx.send("Sorry, you must be admin to use that command!")
+
+
+    async def on_ready(self):
+        print('Logged in as')
+        print(self.user.name)
+        print(self.user.id)
+        print('------')
